Starting to refactor, splitting up files
